+++ cypress/support/commands.js
@@ -24,6 +24,6 @@
 // -- This is will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
 
-Cypress.Commands.add(`assertRoute`, route => {
+Cypress.Commands.add(`assertRoute`, (route) => {
   cy.url().should(`equal`, `${window.location.origin}${route}`)
 })