+++ themes/gatsby-theme-minimal-blog/src/gatsby-plugin-theme-ui/components.js
@@ -12,7 +12,7 @@ export default {
       {children}
     </Title>
   ),
-  pre: preProps => {
+  pre: (preProps) => {
     const props = preToCodeBlock(preProps)
     // if there's a codeString and some props, we passed the test
     if (props) {