+++ themes/gatsby-theme-specimens/src/components/color-swatch.tsx
@@ -58,7 +58,7 @@ const ColorSwatch = ({ color, name, minimal, className, prefix }: ColorSwatchPro
           position: `relative`,
         }}
       >
-        {ratings.map(rating => (
+        {ratings.map((rating) => (
           <div
             key={`${rating.color}-${rating.size}-${rating.value}`}
             sx={{ display: `flex`, flexDirection: `column`, margin: `0 3px`, width: `50px` }}