+++ examples/emma/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/LekoArts/gatsby-themes/compare/emma@0.1.7...emma@0.1.8) (2020-03-31)
+
+**Note:** Version bump only for package emma
+
+
+
+
+
 ## [0.1.7](https://github.com/LekoArts/gatsby-themes/compare/emma@0.1.6...emma@0.1.7) (2020-03-04)
 
 **Note:** Version bump only for package emma