+++ examples/graphql-playground/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.14](https://github.com/LekoArts/gatsby-themes/compare/graphql-playground@0.0.13...graphql-playground@0.0.14) (2020-03-31)
+
+**Note:** Version bump only for package graphql-playground
+
+
+
+
+
 ## [0.0.13](https://github.com/LekoArts/gatsby-themes/compare/graphql-playground@0.0.12...graphql-playground@0.0.13) (2020-03-04)
 
 **Note:** Version bump only for package graphql-playground