+++ examples/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.13](https://github.com/LekoArts/gatsby-themes/compare/styleguide@0.0.12...styleguide@0.0.13) (2020-03-31)
+
+**Note:** Version bump only for package styleguide
+
+
+
+
+
 ## [0.0.12](https://github.com/LekoArts/gatsby-themes/compare/styleguide@0.0.11...styleguide@0.0.12) (2020-03-04)
 
 **Note:** Version bump only for package styleguide